Added menu/text mesh pro package
This is the menu I created. The options menu current doesn't have any functionality, but once we add in more items I can add to it. The background is a filler (I don't know where to get the background file, if there even is one). I also added in the text mesh pro package as this is what I used to format the menu.