Menu and Music implemented
Reconfigured build settings so the game opens on the menu and loads level 1 when "start" is clicked. Also added level 3 (cavern level) which is currently identical to level 1. Now each door leads to the next level until the final door leads back to the menu. Also added music to each level and added the credits for each to the credit screen, completely breaking it.