Added Health Bar and Health Tracking
Added functionality for health and a health bar. Health bar is tied to the camera, isnt quite in the right spot for now, and is only on level 1 for now. Also added a 2 second cooldown on taking damage to prevent you from getting hit 4 times in 4 frames. I picked four health to line up nicely with my hp bar sprites but that could be changed.